Getting Started with Styleguidist
https://create-react-app.dev/docs/developing-components-in-isolation#getting-started-with-styleguidist

Styleguidist combines a style guide, where all your components are presented on a single page with their props documentation and usage examples, with an environment for developing components in isolation, similar to Storybook. In Styleguidist you write examples in Markdown, where each code snippet is rendered as a live editable playground.